extract hashNexusTreeNames to a separate file
